// src/content/pages/MyChefsbase/Ingredients/types/AddIngredient.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

import { AddIngredientInput } from "./../../../../../globalTypes";

// ====================================================
// GraphQL mutation operation: AddIngredient
// ====================================================

export interface AddIngredient {
  addIngredient: boolean | null;
}

export interface AddIngredientVariables {
  input: AddIngredientInput;
  products?: string[] | null;
}